Try-Except for Gene_option error at website lookup
For some reason, when using the glia database a user reported an error when navigating to the Allele and custom-selection tab. It seems to stem from some of the genes/alleles not on the wormbase_ID lookup list (especially N2). I added a try/except statement at lines 230 and 690 to get over error in the meantime.